% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_exchange_PML.R
\name{coord_exchange_PML}
\alias{coord_exchange_PML}
\title{Search for an optimal choice experiment design under the panel mixed logit model}
\usage{
coord_exchange_PML(
  n_choice_set,
  n_alternative,
  n_level_attribute,
  effect_mean,
  effect_vars,
  approx = "MQL",
  n_run = 10,
  optcrit = "D"
)
}
\arguments{
\item{n_choice_set}{Number of choice sets}

\item{n_alternative}{The number of alternatives in each choice set}

\item{n_level_attribute}{Vector where each element is the number of levels for a factor. For example, for two 2-level factors and one 3-level factor this would contain c(2, 2, 3).}

\item{effect_mean}{Vector of means for the effects coded attribute effects}

\item{effect_vars}{Vector of variances for the effects coded attribute effects}

\item{approx}{Type of approximation to use (MQL, PQL, Importance, Laplace, or MSM)}

\item{n_run}{Number of times to run the coordinate exchange algorithm}

\item{optcrit}{Optimality criterion. Either "A" or "D"}
}
\value{
List object containing the following results of the coordinate exchange algorithm: \tabular{ll}{
   \code{design_list} \tab A list object with the final design from each coordinate exchange run \cr
   \tab \cr
   \code{modmat_list} \tab  A list object with the final model matrix from each coordinate exchange run \cr
   \tab \cr
   \code{criteria_list} \tab  A list object with the optimality criteria value for the final design from each coordinate exchange run \cr
   \tab \cr
   \code{best_design} \tab  The best design found by the coordinate exchange algorithm \cr
   \tab \cr
   \code{best_design_varcov} \tab  The variance-covariance matrix for the model parameters for the best design found by the coordinate exchange algorithm. The form will depend on which approximation method was used. \cr
   \tab \cr
   \code{best_model_matrix} \tab  The model matrix corresponding to the best design found by the coordinate exchange algorithm \cr
   \tab \cr
   \code{best_criteria_value} \tab  The value of the optimality criteria for the best design found by the coordinate exchange algorithm \cr
   \tab \cr
   \code{computation_times} \tab  The computation time for each run of the coordinate exchange algorithm \cr
   \tab \cr
   \code{n_accepted_exchange_list} \tab  The number of exchanges accepted during each run of the coordinate exchange algorithm. Can be used to verify that the algorithm is successfully moving away from the initial design \cr
   \tab \cr
   \code{n_proposal_evaled_list} \tab  The number of exchanges proposed during each run of the coordinate exchange algorithm \cr
   \tab \cr
   \code{overall_search_time} \tab  Overall time required to search for a design across all coordinate exchange algorithm runs \cr
}
}
\description{
Search for an optimal choice experiment design under the panel mixed logit model
}
\examples{

# 3 attributes
nattr <- 3

# First two attributes have two levels, third attribute has 3 levels
n_level_attribute <- c(2, 2, 3)

# 2 alternatives per choice set
n_alternative <- 2 

# Total of 8 choice sets
n_choice_set <- 8

# effect means
mu <- c(1.0, -0.4, -0.8, 0.3)

# effect variances
sig <- c(0.1, 0.3, 0.4, 0.4)

result <- coord_exchange_PML(n_choice_set, n_alternative, n_level_attribute, mu, sig, n_run = 10)

}
