% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQL.R
\name{cprob}
\alias{cprob}
\title{Calculate choice probabilities}
\usage{
cprob(b, modmat, qes)
}
\arguments{
\item{b}{Vector of effects}

\item{modmat}{The model matrix}

\item{qes}{A vector of alternative ids (eg 1, 2, 3, 1, 2, 3, 1, 2, 3)}
}
\value{
The probability of each alternative being selected in each choice set
}
\description{
Calculate choice probabilities
}
