% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQL.R
\name{evaluate_choice_probabilities}
\alias{evaluate_choice_probabilities}
\title{Calculate choice probabilities for each alternative in each choice set}
\usage{
evaluate_choice_probabilities(b, model_matrix, alternative_ordering)
}
\arguments{
\item{b}{Vector of effects}

\item{model_matrix}{The model matrix. Dimension should be}

\item{alternative_ordering}{A vector of alternative ids (eg 1, 2, 3, 1, 2, 3, 1, 2, 3)}
}
\value{
The probability of each alternative being selected in each choice set
}
\description{
Calculate choice probabilities for each alternative in each choice set
}
\keyword{internal}
