% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gen_all_choice_seq}
\alias{gen_all_choice_seq}
\title{Generate all possible response sequences for a DCE}
\usage{
gen_all_choice_seq(nChoiceSet, nAlternative)
}
\arguments{
\item{nChoiceSet}{Number of choice sets}

\item{nAlternative}{Number of alternatives per choice set}
}
\value{
An array with all possible response sequences to the experiment
}
\description{
Generate all possible response sequences for a DCE
}
