% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{importanceSampleFixedY}
\alias{importanceSampleFixedY}
\title{Approximate the information matrix using importance sampling}
\usage{
importanceSampleFixedY(Y, X, b_mean, var_vec, nChoiceSet, nU)
}
\arguments{
\item{Y}{an enumeration of all possible responses. Can be obtained using
the \code{\link{gen_all_choice_seq}} function}

\item{X}{The model matrix}

\item{b_mean}{The vector of effect means}

\item{var_vec}{The vector of effect variances}

\item{nChoiceSet}{The number of choice sets in the discrete choice experiment}

\item{nU}{Number of samples to use to evalutate the information matrix}
}
\value{
A matrix containing the MSM evaluated of the variance
covariance matrix of the model parameters
}
\description{
Approximates the information matrix for the model parameters 
using importance sampling. This is not intended to be directly used,
instead a user should use the \code{\link{PMLInfoApprox}} function with 
the "importance" option.
}
