% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{laplaceApproximation}
\alias{laplaceApproximation}
\title{Laplace approximation to the information matrix}
\usage{
laplaceApproximation(ny, X, b_mean, var_vec, n_choice_set)
}
\arguments{
\item{ny}{Number of response samples to draw}

\item{X}{The model matrix}

\item{b_mean}{The vector of effect means}

\item{var_vec}{The vector of effect variances}

\item{n_choice_set}{The number of choice sets in the discrete choice experiment}
}
\value{
A matrix containing the Laplace approximation to the variance covariance matrix of the model parameters
}
\description{
Approximates the information matrix for the model parameters 
using a Laplace Approximation. This is not intended to be directly used,
instead a user should use the \code{\link{PMLInfoApprox}} function with the "Laplace" option.
}
