% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQL.R
\name{MQLApprox}
\alias{MQLApprox}
\title{Marginal Quasi Likelihood approach to approximate the information matrix for the
model parameters}
\usage{
MQLApprox(modmat, effectMean, effectVar, nChoiceSet)
}
\arguments{
\item{modmat}{The model matrix}

\item{effectMean}{Vector of means for the effects coded attribute effects}

\item{effectVar}{Vector of variances for the effects coded attribute effects}

\item{nChoiceSet}{Number of choice sets}
}
\value{
The approximation to the information matrix for the model parameters
}
\description{
Marginal Quasi Likelihood approach to approximate the information matrix for the
model parameters
}
