% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MSM_approx_vcov}
\alias{MSM_approx_vcov}
\title{Approximation to the variance-covariance matrix for the panel mixed logit model using the method of simulated moments}
\usage{
MSM_approx_vcov(n_samples, X, b_mean, var_vec, n_choice_set)
}
\arguments{
\item{n_samples}{Number of response samples to draw}

\item{X}{The model matrix}

\item{b_mean}{The vector of effect means}

\item{var_vec}{The vector of effect variances}

\item{n_choice_set}{The number of choice sets in the discrete choice experiment}
}
\value{
A matrix containing the MSM evaluated approximation of the variance
covariance matrix of the model parameters
}
\description{
Approximates the variance covariance matrix for the model parameters 
using the method of simulated moments. This is not intended to be directly used,
instead a user should use the \code{\link{varcov_approx_PML}} function with 
the "MSM" option.
}
