% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MSMApprox}
\alias{MSMApprox}
\title{Approximation to the information matrix using the method of simulated moments}
\usage{
MSMApprox(ny, X, b_mean, var_vec, nChoiceSet)
}
\arguments{
\item{ny}{Number of response samples to draw}

\item{X}{The model matrix}

\item{b_mean}{The vector of effect means}

\item{var_vec}{The vector of effect variances}

\item{nChoiceSet}{The number of choice sets in the discrete choice experiment}
}
\value{
A matrix containing the MSM evaluated of the variance
covariance matrix of the model parameters
}
\description{
Approximates the information matrix for the model parameters 
using the method of simulated moments. This is not intended to be directly used,
instead a user should use the \code{\link{PMLInfoApprox}} function with 
the "MSM" option.
}
