% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanelMixedDesign-package.R
\docType{package}
\name{PanelMixedDesign-package}
\alias{PanelMixedDesign}
\alias{PanelMixedDesign-package}
\title{PanelMixedDesign: Approximations to the Variance-Covariance Matrix of the Model Parameters for Choice Experiments Modeled Under the Panel Mixed Logit Model}
\description{
Provides several approximations to the variance-covariance matrix under the Panel Mixed Logit model. These approximations are faster to calculate than using numerical integration, which makes them useful for searching for optimal designs for discrete choice experiments. The package also includes a coordinate exchange algorithm for searching for designs for DCEs using the approximations.
}
\keyword{internal}
