% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PQL_approx_vcov}
\alias{PQL_approx_vcov}
\title{Approximation to the information matrix using PQL}
\usage{
PQL_approx_vcov(n_samples, X, b_mean, var_vec, nChoiceSet)
}
\arguments{
\item{n_samples}{Number of samples to draw}

\item{X}{The model matrix}

\item{b_mean}{The vector of effect means}

\item{var_vec}{The vector of effect variances}

\item{nChoiceSet}{The number of choice sets in the discrete choice experiment}
}
\value{
A matrix containing the PQL approximation to the variance
covariance matrix of the model parameters
}
\description{
Approximates the information matrix for the model parameters 
using partial quasilikelihood. This is not intended to be directly used,
instead a user should use the \code{\link{varcov_approx_PML}} function with 
the "PQL" option.
}
\keyword{internal}
