\name{PanelMixedDesign-package}
\alias{PanelMixedDesign-package}
\alias{PanelMixedDesign}
\docType{package}
\title{
\packageTitle{PanelMixedDesign}
}
\description{
\packageDescription{PanelMixedDesign}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{PanelMixedDesign}
\packageIndices{PanelMixedDesign}
Provides serval approximations to the variance-covariance matrix under the Panel Mixed Logit model. These approximations are faster to calculate than using numerical integration, which makes them useful for searching for optimal designs for discrete choice experiments. The package also includes a coordinate exchange algorithm for searching for designs for DCEs using the approximations.
}
\author{
\packageAuthor{PanelMixedDesign}
Maintainer: Joshua Lukemire\packageMaintainer{PanelMixedDesign}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ panel mixed logit}
